Fix macos build

+++ src/features/mm_bone_data_feature.h
@@ -13,7 +13,7 @@ public:
 
     virtual void setup_for_animation(Ref<Animation> animation) override;
 
-    virtual size_t get_dimension_count() const override;
+    virtual uint32_t get_dimension_count() const override;
 
     virtual PackedFloat32Array bake_animation_pose(Ref<Animation> p_animation, float time) const override;
