+++ src/features/mm_trajectory_feature.h
@@ -12,7 +12,7 @@ public:
     MMTrajectoryFeature(/* args */);
     virtual ~MMTrajectoryFeature();
 
-    virtual size_t get_dimension_count() const override;
+    virtual uint32_t get_dimension_count() const override;
 
     virtual void setup_skeleton(const MMAnimationPlayer* p_player, const Skeleton3D* p_skeleton) override;
 
@@ -37,7 +37,7 @@ protected:
     static void _bind_methods();
 
 private:
-    size_t _get_point_dimension_count() const;
+    uint32_t _get_point_dimension_count() const;
 
     int _root_bone{-1};
     NodePath _root_bone_path;
